+++ projects/oblique/src/lib/service-navigation/api/service-navigation.api.model.ts
@@ -124,4 +124,5 @@ export interface ObIServiceNavigationApplicationIdentifier {
 export interface ObIProfileState {
 	avatarID: number;
 	fullname: string;
+	language?: string;
 }